move options to transformer package
